Create House Robber -- Leetcode
Approach -->

The problem can be seen as a decision-making problem, if at each house, whether to rob it or not. Since adjacent houses have security systems connected, robbing one house means the neighboring house cannot be robbed. The goal is to find the maximummaximummaximum amount of money that can be robbed without alerting the police.

To solve this problem, we can use dynamic programming to keep track of the maximum amount of money that can be robbed up to the current house. Let's define an array dp, where dp[i] represents the maximum amount of money that can be robbed up to the i-th house.

The recursive relation for the dynamic programming approach is as follows:
dp[i]=max( dp[i-1] , dp[i-2] + nums[i])

Here, dp[i-1] = represents the maximum amount of money robbed excluding the current house.
dp[i-2] + nums[i] = represents the maximum amount of money robbed including the current house. We choose the maximum of these two options.